cleaning up the dependency chain before restructuring
