+++ draco-oxide/src/decode/attribute/inverse_prediction_transform/oct_reflection.rs
@@ -3,7 +3,7 @@ use std::result::Result;
 use super::InversePredictionTransformImpl;
 use crate::core::shared::{DataValue, NdVector, Vector}; 
 use crate::decode::attribute::portabilization::{Deportabilization, DeportabilizationImpl};
-use crate::encode::attribute::prediction_transform::geom::*;
+use crate::shared::attribute::geom::*;
 use crate::prelude::ByteReader;
 use crate::shared::attribute::Portable;
