+++ draco-oxide/src/encode/attribute/prediction_transform/mod.rs
@@ -1,5 +1,4 @@
 pub mod difference;
-pub(crate) mod geom;
 pub mod oct_orthogonal;
 pub mod oct_reflection;
 pub mod orthogonal;
