+++ draco-oxide/src/encode/entropy/mod.rs
@@ -1,2 +1,1 @@
-pub mod rans;
 pub mod symbol_coding;
